Move field selection map value-node converter to Converters
