+++ src/HotChocolate/Fusion/src/Fusion.Utilities/Rewriters/ValueSelectionToSelectionSetRewriter.cs
@@ -1,3 +1,4 @@
+using HotChocolate.Fusion.Converters;
 using HotChocolate.Fusion.Language;
 using HotChocolate.Language;
 using HotChocolate.Types;
